# [Rust] improve perf