--- a/PrimeRust/solution_1/Cargo.toml~
+++ b/PrimeRust/solution_1/Cargo.toml~
@@ -14,3 +14,4 @@ structopt = "0.3"
 opt-level = 3
 lto = true
 codegen-units = 1
+