@ItalyToast 2021-08-25
I think this change would break faithfulness according to rule 3: 
> The sieve size and corresponding prime candidate memory buffer (or language equivalent) are set/allocated dynamically at runtime. The size of the memory buffer must correspond to the size of the sieve.
Solution owners generally want their solution on the leaderboard.

Tagging @mike-barber for feedback.