@pro465 2021-08-25
@ItalyToast but the nim solution no. 3 has also marked the table as `const` and is also marked as faithful 🤔