@ItalyToast 2021-08-25
> @ItalyToast but the nim solution no. 3 has marked the corresponding map/table as `const` and is also marked as faithful 🤔

My bad, I'm not very proficient in rust, I assumed you made a change so the program could only compute a 1 million sized sieve. The validation map is of course allowed to be a constant.