@ItalyToast 2021-08-25
I think it would be nice if you could provide an approximation of the increase you see in performance. 

> soo...no problem i guess?
> 
> anyways i still have to get approval from a maintainer. could you tell me where to look for? i am very new to github

In this repository you first need to get approval from the solution owner which is mike in this case and when he have approved the changes a maintainer will look at the code and make sure it conforms to the rules. I have already tagged mike for you so he'll probably get back to you within a day(it might take a little bit longer ofc), hes usually very active.