@mike-barber 2021-08-25
I will need to validate this before giving it the OK from my side. It needs to have a measurable performance improvement for it to make sense merging. 

Can you confirm what performance improvement you're seeing -- before and after numbers, and the CPU you're on? 

Thanks
M