@mike-barber 2021-08-25
That is very cool! I'm really glad to hear that you're able to run Rust on Android, and you're not blocked by not having access to a PC. 

Thanks for the suggestions in the PR. I have run them on both my main workstation and a Raspberry Pi 4 (which also actually works quite well for playing around with Rust). 

Unfortunately, it looks like there are no performance improvements from the PR. It looks very slightly _slower_ on the Raspberry Pi, but it's very close to being exactly the same. 

Here are the numbers I got -- I did 10 repeated runs to get more accurate results, and just the fastest solution we're interested in improving; you can do this too with `cargo run --release -- -t 1 -r 10 --bits-striped-hybrid 2> /dev/null` but I suspect your device will get quite hot and slow down pretty quickly doing a bunch of repeated runs.

## Computer - Ryzen 3900X 

### original

```
mike-barber_bit-storage-striped-hybrid;26657;5.0001649857;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26728;5.0000834465;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26676;5.0001754761;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26641;5.0000691414;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26642;5.0000824928;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26744;5.0001840591;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26681;5.0002279282;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26668;5.0001702309;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26708;5.0001182556;1;algorithm=base,faithful=yes,bits=1
```

### PR 

```
mike-barber_bit-storage-striped-hybrid;26612;5.0000615120;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26627;5.0002021790;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26693;5.0002317429;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26432;5.0001039505;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26513;5.0001220703;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26681;5.0001139641;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26678;5.0000729561;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26729;5.0000529289;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26680;5.0002140999;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid;26675;5.0001544952;1;algorithm=base,faithful=yes,bits=1
```

measurement (normal blocks) | original | pr | ratio 
----|---|---|---
total iterations	| 266849 | 266320 | 
total time | 50.0013895 | 50.0013299 | 
iterations / 5s | 26684.15845	| 26631.29166 | 0.998018795 




## Raspberry Pi 4 

We're interested in the small blocks on this processor, as we have less cache.

### original

```
mike-barber_bit-storage-striped-hybrid-small;3860;5.0002708435;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3859;5.0012722015;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3864;5.0007920265;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3861;5.0010852814;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3861;5.0011577606;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3863;5.0008997917;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3861;5.0010910034;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3861;5.0012197495;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3861;5.0010848045;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3861;5.0003552437;1;algorithm=base,faithful=yes,bits=1
```

### PR

```
mike-barber_bit-storage-striped-hybrid-small;3829;5.0002884865;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3830;5.0007376671;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3833;5.0008640289;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3828;5.0002326965;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3830;5.0008482933;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3829;5.0012712479;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3830;5.0007848740;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3829;5.0005941391;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3830;5.0004920959;1;algorithm=base,faithful=yes,bits=1
mike-barber_bit-storage-striped-hybrid-small;3828;5.0012502670;1;algorithm=base,faithful=yes,bits=1
```

measurement (small blocks) | original | pr | ratio 
----|---|---|---
total iterations	| 38612 | 38296 | 
total time | 50.00922871 | 50.0073638 | 
iterations / 5s | 3860.487454	| 3829.036075 | 0.991853003


### Conclusions

Overall, it looks like the changes in the PR make the solution very slightly slower on average, but only by about 0.81%. I think I know why.
- the changes to the validator will make no difference, as it's only used for testing and reporting at the end of the run -- it's not in the timed loop.
- the mask generation might be slightly less efficient, as noted below.

```rust
let mut mask_set = [[0u8; U8_BITS]; SKIP];
for (word_idx, word) in mask_set.iter_mut().enumerate().take(SKIP) {
    for (bit_idx, bit) in word.iter_mut().enumerate().take(8) {
        let block_index_offset = block_idx * BLOCK_SIZE * U8_BITS;
        let bit_index_offset = bit_idx * BLOCK_SIZE;
        let index = block_index_offset + bit_index_offset + word_idx;
        *bit = !(Self::should_reset(index, start, SKIP) << bit_idx);
    }
}
```
Normally, using iterators would be a good idea, especially with vectors. Direct indexing is really expensive due to the bounds checks. 

However, `mask` has dimensions known at compile time, so it's perfectly fine addressing elements directly: the compiler will elide bounds checks at compile time. Recall that `U8_BITS` and `SKIP` are constants.

```rust
let mut mask_set = [[0u8; U8_BITS]; SKIP];
for word_idx in 0..SKIP {
    for bit in 0..8 {
        let block_index_offset = block_idx * BLOCK_SIZE * U8_BITS;
        let bit_index_offset = bit * BLOCK_SIZE;
        let index = block_index_offset + bit_index_offset + word_idx;
        mask_set[word_idx][bit] = !(Self::should_reset(index, start, SKIP) << bit);
    }
}
```

In this particular case, it looks like the iterator is very slightly slower for some reason. It might be due to the `take()`, but I can't see anything obvious in the assembly. It's also worth noting that the time spend initialising the masks is almost zero compared to the overall time spent resetting the bits in the main loop, so there's very little to be gained from improving it. 

Unfortunately, because this PR doesn't measurably improve performance, I'm going to ask for us to close it. But thanks again for taking the time to play around with the solution and try to improve it. I hope my explanation and justifications make sense. I really appreciate your interest and input!